@@ -1,17 +1,15 @@
-﻿namespace AddRemoveProgramsCleaner.Registry {
+﻿namespace AddRemoveProgramsCleaner.Registry; 
 
-    public interface RegistryConstants {
+public interface RegistryConstants {
 
-        const string DISPLAY_NAME        = "DisplayName";
-        const string DISPLAY_ICON        = "DisplayIcon";
-        const string SYSTEM_COMPONENT    = "SystemComponent";
-        const string INSTALL_LOCATION    = "InstallLocation";
-        const string UNINSTALL_STRING    = "UninstallString";
-        const string PRODUCT_NAME        = "ProductName";
-        const string PRODUCT_NAME_HIDDEN = "ProductName hidden by Ben";
-        const string PRODUCT_ICON        = "ProductIcon";
-        const string RELEASE_TYPE        = "ReleaseType";
-
-    }
+    const string DISPLAY_NAME        = "DisplayName";
+    const string DISPLAY_ICON        = "DisplayIcon";
+    const string SYSTEM_COMPONENT    = "SystemComponent";
+    const string INSTALL_LOCATION    = "InstallLocation";
+    const string UNINSTALL_STRING    = "UninstallString";
+    const string PRODUCT_NAME        = "ProductName";
+    const string PRODUCT_NAME_HIDDEN = "ProductName hidden by Ben";
+    const string PRODUCT_ICON        = "ProductIcon";
+    const string RELEASE_TYPE        = "ReleaseType";
 
 }
